requirements.txt
streamlit
opencv-python
numpy
pandas
scikit-learn
joblib
kagglehub
opencv-contrib-python
Pillow